Moved types into separate files. Feature completed fixes #6
