+++ Source/AsyncGenerator.Core/Analyzation/IAccessorAnalyzationResult.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis;
+
+namespace AsyncGenerator.Core.Analyzation
+{
+	public interface IAccessorAnalyzationResult : IMethodOrAccessorAnalyzationResult
+	{
+		SyntaxNode Node { get; }
+	}
+}
