+++ Source/AsyncGenerator.Core/Transformation/IAccessorTransformationResult.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using AsyncGenerator.Core.Analyzation;
+
+namespace AsyncGenerator.Core.Transformation
+{
+	public interface IAccessorTransformationResult : IMethodOrAccessorTransformationResult
+	{
+		new IAccessorAnalyzationResult AnalyzationResult { get; }
+	}
+}
