+++ Source/AsyncGenerator.Core/Transformation/IMethodTransformationResult.cs
@@ -10,25 +10,4 @@ namespace AsyncGenerator.Core.Transformation
 		new IMethodAnalyzationResult AnalyzationResult { get; }
 
 	}
-
-	public interface IAccessorTransformationResult : IMethodOrAccessorTransformationResult
-	{
-		new IAccessorAnalyzationResult AnalyzationResult { get; }
-	}
-
-	public interface IMethodOrAccessorTransformationResult : IMemberTransformationResult
-	{
-		/// <summary>
-		/// The transformed method
-		/// </summary>
-		MethodDeclarationSyntax Transformed { get; }
-
-		IMethodOrAccessorAnalyzationResult AnalyzationResult { get; }
-
-		SyntaxTrivia BodyLeadingWhitespaceTrivia { get; }
-
-		IReadOnlyList<IFunctionReferenceTransformationResult> TransformedFunctionReferences { get; }
-
-		IReadOnlyList<ILockTransformationResult> TransformedLocks { get; }
-	}
 }
