+++ Source/AsyncGenerator/Transformation/Internal/CancellationTokenMethodTransformer.cs
@@ -20,7 +20,7 @@ using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
 namespace AsyncGenerator.Transformation.Internal
 {
-	public class CancellationTokenMethodTransformer : IMethodOrAccessorTransformer
+	internal class CancellationTokenMethodTransformer : IMethodOrAccessorTransformer
 	{
 		private IProjectCancellationTokenConfiguration _configuration;
