+++ Source/AsyncGenerator/Transformation/Internal/YieldMethodTransformer.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using AsyncGenerator.Core.Configuration;
+using AsyncGenerator.Core.Plugins;
+using AsyncGenerator.Core.Transformation;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace AsyncGenerator.Transformation.Internal
+{
+	internal class YieldMethodTransformer : IMethodOrAccessorTransformer
+	{
+		public Task Initialize(Project project, IProjectConfiguration configuration)
+		{
+			return Task.CompletedTask;
+		}
+
+		public MethodTransformerResult Transform(IMethodOrAccessorTransformationResult transformResult,
+			ITypeTransformationMetadata typeMetadata, INamespaceTransformationMetadata namespaceMetadata)
+		{
+			var methodResult = transformResult.AnalyzationResult;
+			if (!methodResult.RewriteYields)
+			{
+				return MethodTransformerResult.Skip;
+			}
+			var methodNode = transformResult.Transformed;
+			var rewriter = new YieldRewriter(transformResult);
+			methodNode = (MethodDeclarationSyntax)rewriter.VisitMethodDeclaration(methodNode);
+			return MethodTransformerResult.Update(methodNode);
+		}
+	}
+}
