+++ Source/AsyncGenerator/Transformation/Internal/YieldRewriter.cs
@@ -14,28 +14,6 @@ using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
 namespace AsyncGenerator.Transformation.Internal
 {
-	internal class YieldTransformer : IMethodOrAccessorTransformer
-	{
-		public Task Initialize(Project project, IProjectConfiguration configuration)
-		{
-			return Task.CompletedTask;
-		}
-
-		public MethodTransformerResult Transform(IMethodOrAccessorTransformationResult transformResult,
-			ITypeTransformationMetadata typeMetadata, INamespaceTransformationMetadata namespaceMetadata)
-		{
-			var methodResult = transformResult.AnalyzationResult;
-			if (!methodResult.RewriteYields)
-			{
-				return MethodTransformerResult.Skip;
-			}
-			var methodNode = transformResult.Transformed;
-			var rewriter = new YieldRewriter(transformResult);
-			methodNode = (MethodDeclarationSyntax)rewriter.VisitMethodDeclaration(methodNode);
-			return MethodTransformerResult.Update(methodNode);
-		}
-	}
-
 	internal class YieldRewriter : CSharpSyntaxRewriter
 	{
 		private readonly IMethodOrAccessorTransformationResult _transformResult;
